feat(auth): Add back AuthAppMiddleware.from_app()
This makes it more convenient to create a middleware instance from
within a flask app context.

A conditional import is used to not create a hard dependency on flask.
The method’s behavior is undefined for when flask is not installed so
it’s OK to not have tests for this.